lustre: change subdir-ccflags-y to ccflags-y
Using subdir-ccflags-y imposes the flags on all
subdirectories - but only when you make from a high level.
If you "make directory/name/" it doesn't find
the subdir-ccflags-y from parent Makefiles.

Because of this weakness, lustre includes the same setting
in every Makefile (that compiles code).
Given this, we may as well just use ccflags-y.  This makes
the purpose more obvious, and is a much more widely used macro.