staging: lustre: selftest: fix time reporting
Migrate crp_stamp to nanoseconds and both timestamps nd_stamp,
sn_start to ktime. The move away from jiffies, which can vary on
platforms, to something that is the consistent on every node.
Currently jiffies data was being transmitted over the wire and
the other node could have a different HZ value configured which
lead to incorrect results. This will ensure that the reported
results to the user will always be correct.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
WC-bug-id: https://jira.whamcloud.com/browse/LU-9019
Reviewed-on: https://review.whamcloud.com/31041
Reviewed-by: Doug Oucharek <dougso@me.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>