lustre: simplify use of interval-tree.
Lustre has a private interval-tree implementation.  This
implementation refuses to insert an interval if an
identical interval already exists.  It is OK with all sorts of
overlapping intervals, but identical intervals are rejected.
This functionality is used in the server-side code, but not
in the client.

Both users of interval-tree in lustre would be simpler if this was not
the case.  They need to store all intervals, even if some are
identical.

llite/range_lock.c adds a rl_next_lock list_head to each lock.
If it cannot insert a new lock because the range is in use, it
attached the new lock to the existing lock using rl_next_lock.
This requires extra code to iterate over the rl_next_lock lists when
iterating over locks, and to update the list when deleting a lock from
the tree.

ldlm_extend allocates a separate ldlm_interval which as a list of
ldlm_locks which share the same interval.  This is linked together
by over-loading the l_sl_policy which, for non-extent locks, is used
for linking together locks with the same policy.
This doesn't only require extra code, but also an extra memory
allocation.

This patch removes all that complexity.
- interval_insert() now never fails.
- consequently rl_next_lock is always empty and
  rl_lock_count is always zero. so they are removed
- every ldlm_lock has linked directly into the
  interval tree, so each has an embedded interval_node
  rather than a pointer to a 'struct ldlm_interval'
- ldlm_interval is now unused, so it is gone as it
  the kmemcache from which they were allocated.
- the various functions for allocating an ldlm_interval
  and attaching to a lock or detaching from a lock
  are also gone.

The server-side users could easily be changed to perform a lookup
first if they want to ensure there are no duplicates.

Signed-off-by: NeilBrown <neilb@suse.com>