lustre: llite: replace several GFP_NOFS with GFP_KERNEL
GFP_KERNEL should be used preferentially unless there is
good reason to use something else.

GFP_NOFS is appropriate when the code holds some resource
that could blocks writeout triggered by memory reclaim.
So it is appropriate to us it in the writeout path, or anything
that could interact with the writeout paths.

The memory allocations affected here are either ioctl, or open/close
handling.  These can never block writeout, so GFP_NOFS is not needed.

Signed-off-by: NeilBrown <neilb@suse.com>