lustre: remove dead code from obd_connect_flags2str()
In obd_connect_flags2str(), if obd_connect_names() has
fewer than 64 non-NULL entries, then the first for-loop
is dangerous.
If it has 64 or more, the first "unknown" error is pointless.

In fact, there are more than 64 entries, so provide a BUILD_BUG_ON
to that effect, and remove the pointless code.

Signed-off-by: NeilBrown <neilb@suse.com>