lustre: fix inode refcount problem with fhandle access
ll_iget_for_nfs() currently calls d_obtain_alias() twice.
This decrements the refcount on the inode twice, and
increments it on the dentry twice.
In each case, only once is correct.
The result is that the refcounts are wrong and warnings
result.

Also, if the first d_obtain_alias() returns an error,
iput() is called.  As the comment at the second d_obtain_alias()
explains, this is incorrect.

Reported-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Fixes: c1b66fccf986 ("staging: lustre: fid: do open-by-fid by default")
Signed-off-by: NeilBrown <neilb@suse.com>