lustre: lnet/config: convert list_for_each to list_for_each_entry
This conversion to list_for_each_entry isn't quite trivial
as the 'tmp' loop variables are used elsewhere in the function.
This means we cannot just delete them, and so must inspect the
code to ensure the values aren't used elsewhere - they aren't.

Also we need to introduce new loop variables: ltb1 and ltb2 as
the lnet_text_buf was the some for both the original (nested) loops.

Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>