lustre: convert list_for_each in ksocknal_create_routes
The two list_for_each() loops in ksocknal_create_routes() cannot be
trivially converted to list_for_each_entry() as the new loop variable
is expected to be NULL after the loop, if the loop completed normally.

If the loop did not complete normally - if an interesting entry was
found - a 'continue' is required in the higher level loop.  Change
this 'continue' to a 'goto' so it can be called in the inner loop, and
we don't need to check for a variable being NULL.

Reviewed-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>