lustre: give different tcd_lock types different classes.
There are three different trace contexts:
 process, softirq, irq.
Each has its own lock (tcd_lock) which is locked
as appropriate for that context.
lockdep currently doesn't see that they are different
and so deduces that the different uses might lead to
deadlocks.
So use separate calls to spin_lock_init() so that they
each get a separate lock class, and lockdep sees no
problem.

Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>