lustre/libcfs: discard cfs_trace_allocate_string_buffer()
cfs_trace_allocate_string_buffer() is a simple wrapper
around kzalloc() that adds little value.  The code is
clearer if we perform the test and the allocation
directly where needed.

Also change the test from '>' to '>=' to ensure we
never try to allocate more than 2 pages, as that seems
to be the intent.

Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>