lustre: mgc: relate sptlrpc & param to MGC
If sptlrpc or params config logs come from different MGC,
it should be regarded as different logs, this patch binds
these config logs with MGC obd device to separate them.

The fix for a bug discovered later is also included for
this patch. Since sb is NULL for config_log_find_or_add
the cfs_instance field was being set to the obd device.
This confused the sptlrpc layer so for now cfg_instance
is set to NULl in the sptlrpc case. This will be resolved
with the move to kobjects.

Signed-off-by: Hongchao Zhang <hongchao@whamcloud.com>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: John L. Hammond <jhammond@whamcloud.com>
Intel-bug-id: https://jira.whamcloud.com/browse/LU-9034
Reviewed-on: https://review.whamcloud.com/24988
Intel-bug-id: https://jira.whamcloud.com/browse/LU-9567
Reviewed-on: https://review.whamcloud.com/27320
Reviewed-by: Mike Pershin <mpershin@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>