lustre: llite: rename ldebugfs_[un]register_mountpoint
Rename ldebugfs_[un]register_mountpoint() to
ll_debugfs_[un]register_super(). This first
aligns the functions name with the major of
the llite function prototypes by starting
with the ll_* prefix. The use of mountpoint
is not the best way to describe the functions
who's name are being changed. Since this is
managed by struct ll_sb_info, which is the
lustre private instance for struct super_block,
so rename to *_super() instead.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
WC-bug-id: https://jira.whamcloud.com/browse/LU-8066
Reviewed-on: https://review.whamcloud.com/24031
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>

Changelog)
v1) Initial patch
v2) Improved changelog message
Signed-off-by: NeilBrown <neilb@suse.com>