lustre: llite: remove ll_sb
The only user of ll_sb in struct ll_sb_info is used to query the
name locate in the file_system_type.  That information is
now constant (always "lustre") so are can use a constant
instead.
This enables us to use struct ll_sb_info directly for every sysfs or
debugfs entry.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
WC-bug-id: https://jira.whamcloud.com/browse/LU-8066
Reviewed-on: https://review.whamcloud.com/24031
WC-bug-id: https://jira.whamcloud.com/browse/LU-10824
Reviewed-on: https://review.whamcloud.com/33025
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>