lustre: obd: merge both top lustre sysfs attributes
The top of the lustre sysfs tree is handled by two sets of
attributes, one in sysctl.c and the other in modules.c.
Worst is the the sysfs attributes in sysctl.c are never
cleaned up. So merge all the sysfs handling in sysctl.c
into the one managed in modules.c. Additionally now all
the top level attributes will be cleaned up. This eliminates
the sysctl.c file as well.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
WC-bug-id: https://jira.whamcloud.com/browse/LU-9431
Reviewed-on: https://review.whamcloud.com/30143
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Ben Evans <bevans@cray.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>