lustre: obd: enhance print_lustre_cfg()
Make the function print_lustre_cfg() exportables so we can provide
details when debugging other parts of the stack that handle
lustre_cfg processing. We currently report the index of the buffer
in struct lustre_cfg as well as the buffers length. We don't report
the actually string so lets add that info as well.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
WC-id: https://jira.whamcloud.com/browse/LU-7004
Reviewed-on: https://review.whamcloud.com/28590
Reviewed-by: Fan Yong <fan.yong@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Signed-off-by: NeilBrown <neilb@suse.com>