lnet: store separate xmit/recv net-interface in each message.
Currently we store the net-interface in the peer, but the
peer should identify just the network, not the particular interface.
To help track which actual interface is used for each
message, store them explicitly.

This is part of
    Commit: 8cbb8cd3e771 ("LU-7734 lnet: Multi-Rail local NI split")
from upstream lustre, where it is marked:
    Signed-off-by: Amir Shehata <amir.shehata@intel.com>
    WC-bug-id: https://jira.whamcloud.com/browse/LU-7734
    Reviewed-on: http://review.whamcloud.com/18274
    Reviewed-by: Doug Oucharek <doug.s.oucharek@intel.com>
    Reviewed-by: Olaf Weber <olaf@sgi.com>

and includes commit 63c3e5129873 ("LU-7734 lnet: Fix lnet_msg_free()")

Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Doug Oucharek <dougso@me.com>
Signed-off-by: NeilBrown <neilb@suse.com>